Add dockerfile and fix compilation issues

+++ eve-py/config.sh
@@ -1,19 +1,14 @@
 #!/bin/sh
-cd ltl2ba
-make clean
-make
-cd ..
-if [ -d "temp" ]; then
-	rm -rf temp
-	mkdir temp
-else
-	mkdir temp
-fi
-if [ -d "pgsolver-master" ]; then
-	rm -rf pgsolver-master
-fi
+set -eu
+
+(cd ltl2ba && make clean && make)
+
+[ ! -d temp ] || rm -rf temp
+[ ! -d pgsolver-master ] || rm -rf pgsolver-master
+mkdir -p temp
+
 unzip pgsolver-master
-eval `opam config env`
+eval $(opam config env)
 opam install ocamlbuild ocamlfind ounit TCSLib extlib ocaml-sat-solvers minisat
 cd pgsolver-master
 echo "A" | ocaml setup.ml -configure
